Prevent machine explosion for Universal fluid engine
Currently this machine takes in fluids and then outputs power based  on the current input. When the world is loading in this machine will take in fluids then consume all of it and exploding due to over amping.

Therefore this PR will remove the function of the dynamo exploding and instead the engine will just void all excessive power generation while also consuming the fuel.